… sim/team-validator.ts