In short, dexsearch ~is~ was abnormally slow. A simple query like `/nds fire` took **~2 seconds** and `/nds pivot` **~4 seconds** on the prod server and my decent laptop.

This PR applies a few extremely simple fixes that address ~99% of the performance problems. **The aforementioned queries now take ~5 milliseconds and 30 milliseconds**, respectively. Far more optimizations are possible, but won't be as simple or impactful.

## Results
Below are the (very scarce) performance stats for `npm test`, running on an `i7-9750h`.
These are cheaper than `/nds`. The first test is especially slow because of the nature of JITs.

<details>
<summary>Initial</summary>

```
Before
		now executing: pivot|batonpass, mod=gen8
query prep:	 54.37550900131464 ms
dex size: 804
filters total:	 129.14655699953437 ms
	stat filters:	 0.17849896475672722 ms
	egg filters:	 0.14566901698708534 ms
	type filters:	 0.1426999494433403 ms
	resist filters:	 0.13550499826669693 ms
	weak filters:	 0.14920806139707565 ms
	ability filters: 0.12753695249557495 ms
	forme filters:	 0.12921300157904625 ms
	prepare move filters: 80.43997703120112 ms
		format:   	 74.51695001125336 ms
		ruletable:	 2.823194034397602 ms
		validator:	 2.0557279251515865 ms
		pokesource:	 0.41381001472473145 ms
	move filters:	 45.508391946554184 ms
sort:	 0.6476159989833832 ms
1421 number of unique mon objects from Dex.mod(...)
804 number of above objects that pass gen filters
total run time:			 232.11053999885917 ms
․
		now executing: pivot, mod=gen8
query prep:	 0.7473939992487431 ms
dex size: 804
filters total:	 94.04799500107765 ms
	stat filters:	 0.08011595532298088 ms
	egg filters:	 0.10512598976492882 ms
	type filters:	 0.10283003747463226 ms
	resist filters:	 0.08430599421262741 ms
	weak filters:	 0.11044797673821449 ms
	ability filters: 0.08263105899095535 ms
	forme filters:	 0.0876309834420681 ms
	prepare move filters: 75.64411291107535 ms
		format:   	 71.8328619338572 ms
		ruletable:	 1.4543550126254559 ms
		validator:	 1.648287057876587 ms
		pokesource:	 0.2687230445444584 ms
	move filters:	 16.034791000187397 ms
sort:	 0.38658300042152405 ms
1421 number of unique mon objects from Dex.mod(...)
804 number of above objects that pass gen filters
total run time:			 96.38358999788761 ms
․․․․․․․․․․․
		now executing: ice, monotype
query prep:	 0.1216839998960495 ms
dex size: 873
filters total:	 494.32178499922156 ms
	stat filters:	 0.10472292080521584 ms
	egg filters:	 0.15022199600934982 ms
	type filters:	 0.8074910305440426 ms
	resist filters:	 0.12324101477861404 ms
	weak filters:	 0.11295903846621513 ms
	ability filters: 0.10627801716327667 ms
	forme filters:	 0.10835202410817146 ms
	prepare move filters: 490.29630206152797 ms
		format:   	 76.94467301294208 ms
		ruletable:	 209.55591402947903 ms
		validator:	 202.62786000221968 ms
		pokesource:	 0.504029992967844 ms
	move filters:	 0.5825789906084538 ms
sort:	 0.08564000204205513 ms
2842 number of unique mon objects from Dex.mod(...)
1677 number of above objects that pass gen filters
total run time:			 496.44836499914527 ms
․
		now executing: ice, monotype, spe desc
query prep:	 0.1344659999012947 ms
dex size: 873
filters total:	 444.82719799876213 ms
	stat filters:	 0.09688697010278702 ms
	egg filters:	 0.1446789614856243 ms
	type filters:	 0.5860809981822968 ms
	resist filters:	 0.11443403363227844 ms
	weak filters:	 0.10183602198958397 ms
	ability filters: 0.10196587443351746 ms
	forme filters:	 0.09404192119836807 ms
	prepare move filters: 441.19286999478936 ms
		format:   	 71.15053798630834 ms
		ruletable:	 185.74913900345564 ms
		validator:	 183.20303198322654 ms
		pokesource:	 0.4670259468257427 ms
	move filters:	 0.5626690238714218 ms
sort:	 0.09177299961447716 ms
2842 number of unique mon objects from Dex.mod(...)
1677 number of above objects that pass gen filters
total run time:			 446.4883599989116 ms

		now executing: ice, monotype, hp desc
query prep:	 0.11399000138044357 ms
dex size: 873
filters total:	 446.07669800147414 ms
	stat filters:	 0.10812501609325409 ms
	egg filters:	 0.14091401174664497 ms
	type filters:	 0.6019999720156193 ms
	resist filters:	 0.1115570142865181 ms
	weak filters:	 0.10407200083136559 ms
	ability filters: 0.11563794314861298 ms
	forme filters:	 0.10732902958989143 ms
	prepare move filters: 442.34754206985235 ms
		format:   	 70.71124800294638 ms
		ruletable:	 187.70725399255753 ms
		validator:	 182.57627097144723 ms
		pokesource:	 0.4813690446317196 ms
	move filters:	 0.5883849188685417 ms
sort:	 0.05595799908041954 ms
2842 number of unique mon objects from Dex.mod(...)
1677 number of above objects that pass gen filters
total run time:			 447.76308299973607 ms
```
</details>

<details>
<summary>Latest</summary>

```
After
		now executing: pivot|batonpass, mod=gen8
query prep:	 51.89018200337887 ms
filters total:	 42.756921999156475 ms
sort:		 0.6007909998297691 ms
format result:	 0.09133599698543549 ms
*** entire run time:			 141.8944509997964 ms ***
*** unaccounted:			 46.55522000044584 ms ***
․
		now executing: pivot, mod=gen8
query prep:	 0.6547169983386993 ms
filters total:	 18.113839998841286 ms
sort:		 0.4578099995851517 ms
format result:	 0.06409899890422821 ms
*** entire run time:			 20.0837190002203 ms ***
*** unaccounted:			 0.7932530045509338 ms ***
․․․․․․․․․․․
		now executing: ice, monotype
query prep:	 0.13084299862384796 ms
filters total:	 0.772537000477314 ms
sort:		 0.04282499849796295 ms
format result:	 0.018077000975608826 ms
*** entire run time:			 1.8486009985208511 ms ***
*** unaccounted:			 0.8843189999461174 ms ***
․
		now executing: ice, monotype, spe desc
query prep:	 0.07613600045442581 ms
filters total:	 0.6047910004854202 ms
sort:		 0.03371300548315048 ms
format result:	 0.08656799793243408 ms
*** entire run time:			 1.2767650038003922 ms ***
*** unaccounted:			 0.47555699944496155 ms ***

		now executing: ice, monotype, hp desc
query prep:	 0.055977001786231995 ms
filters total:	 0.6282810047268867 ms
sort:		 0.02927199751138687 ms
format result:	 0.04599300026893616 ms
*** entire run time:			 1.1314250007271767 ms ***
*** unaccounted:			 0.3719019964337349 ms ***
```

</details>

(I go by 'ThereRNoNamesLeft' on Showdown)