@larry-the-table-guy 2024-09-08
Perf comparison for latest (minor) change.
<details>
<summary>Before</summary>

```
Before
		now executing: pivot|batonpass, mod=gen8
query prep:	 56.680244006216526 ms
filtered dex size: 804
filters total:	 52.44317799806595 ms
	stat filters:	 0.11530093848705292 ms
	egg filters:	 0.1262970045208931 ms
	type filters:	 0.11339988559484482 ms
	resist filters:	 0.1280319094657898 ms
	weak filters:	 0.12095899134874344 ms
	ability filters: 0.11294902861118317 ms
	forme filters:	 0.22467390447854996 ms
	prepare move filters: 1.363936997950077 ms
		format:   	 0.21392299979925156 ms
		ruletable:	 0.989206001162529 ms
		validator:	 0.08005800098180771 ms
		pokesource:	 0.07034800201654434 ms
		unaccounted:	 0.010401993989944458 ms
	move filters:	 46.0352219119668 ms
		get moves:	 2.3854240104556084 ms
		learn check:	 42.2000809982419 ms
		unaccounted:	 1.449716903269291 ms
sort 0.7216309979557991 ms
1421 number of unique mon objects from Dex.mod(...)
804 number of above objects that pass gen filters
*** entire run time:			 155.67729900032282 ms ***
*** unaccounted:			 45.832245998084545 ms ***
․
		now executing: pivot, mod=gen8
query prep:	 0.7357559949159622 ms
filtered dex size: 804
filters total:	 23.192718997597694 ms
	stat filters:	 0.08050889521837234 ms
	egg filters:	 0.08070703595876694 ms
	type filters:	 0.07748296111822128 ms
	resist filters:	 0.084907166659832 ms
	weak filters:	 0.08104311674833298 ms
	ability filters: 0.08629320561885834 ms
	forme filters:	 0.07892215996980667 ms
	prepare move filters: 0.16614200174808502 ms
		format:   	 0.14658299833536148 ms
		ruletable:	 0.0077700018882751465 ms
		validator:	 0.009269997477531433 ms
		pokesource:	 0.0016029998660087585 ms
		unaccounted:	 0.0009160041809082031 ms
	move filters:	 20.848288983106613 ms
		get moves:	 1.6765490919351578 ms
		learn check:	 18.15851978212595 ms
		unaccounted:	 1.0132201090455055 ms
sort 0.5495510026812553 ms
1421 number of unique mon objects from Dex.mod(...)
804 number of above objects that pass gen filters
*** entire run time:			 25.92569799721241 ms ***
*** unaccounted:			 1.447672002017498 ms ***
․․․․․․․․․․․
		now executing: ice, monotype
query prep:	 0.11871899664402008 ms
filtered dex size: 873
filters total:	 3.2516840025782585 ms
	stat filters:	 0.07280304282903671 ms
	egg filters:	 0.0771578997373581 ms
	type filters:	 0.35424505919218063 ms
	resist filters:	 0.07407596707344055 ms
	weak filters:	 0.07325991988182068 ms
	ability filters: 0.072704017162323 ms
	forme filters:	 0.07360992580652237 ms
	prepare move filters: 1.4029119983315468 ms
		format:   	 0.09722200036048889 ms
		ruletable:	 0.8441689983010292 ms
		validator:	 0.45731599628925323 ms
		pokesource:	 0.0015619993209838867 ms
		unaccounted:	 0.002643004059791565 ms
	move filters:	 0.23265302926301956 ms
		get moves:	 0.10683392733335495 ms
		learn check:	 0 ms
		unaccounted:	 0.1258191019296646 ms
sort 0.05193500220775604 ms
2842 number of unique mon objects from Dex.mod(...)
1677 number of above objects that pass gen filters
*** entire run time:			 5.17042999714613 ms ***
*** unaccounted:			 1.748091995716095 ms ***
․
		now executing: ice, monotype, spe desc
query prep:	 0.09177500009536743 ms
filtered dex size: 873
filters total:	 2.747948996722698 ms
	stat filters:	 0.07234396040439606 ms
	egg filters:	 0.07728186994791031 ms
	type filters:	 0.2567209228873253 ms
	resist filters:	 0.07246101647615433 ms
	weak filters:	 0.07624068111181259 ms
	ability filters: 0.07200000435113907 ms
	forme filters:	 0.07326596975326538 ms
	prepare move filters: 1.0142109990119934 ms
		format:   	 0.09477400034666061 ms
		ruletable:	 0.5064290016889572 ms
		validator:	 0.4107019975781441 ms
		pokesource:	 0.0013450011610984802 ms
		unaccounted:	 0.0009609982371330261 ms
	move filters:	 0.22153691202402115 ms
		get moves:	 0.10544388741254807 ms
		learn check:	 0 ms
		unaccounted:	 0.11609302461147308 ms
sort 0.045048996806144714 ms
2842 number of unique mon objects from Dex.mod(...)
1677 number of above objects that pass gen filters
*** entire run time:			 4.098100997507572 ms ***
*** unaccounted:			 1.2133280038833618 ms ***

		now executing: ice, monotype, hp desc
query prep:	 0.07578599452972412 ms
filtered dex size: 873
filters total:	 3.006783001124859 ms
	stat filters:	 0.07384903728961945 ms
	egg filters:	 0.09440014511346817 ms
	type filters:	 0.373492032289505 ms
	resist filters:	 0.07403897494077682 ms
	weak filters:	 0.07448684424161911 ms
	ability filters: 0.07351402193307877 ms
	forme filters:	 0.08078712970018387 ms
	prepare move filters: 1.0542949959635735 ms
		format:   	 0.09621100127696991 ms
		ruletable:	 0.516185000538826 ms
		validator:	 0.43790699541568756 ms
		pokesource:	 0.002696998417377472 ms
		unaccounted:	 0.0012950003147125244 ms
	move filters:	 0.23256384581327438 ms
		get moves:	 0.11144090443849564 ms
		learn check:	 0 ms
		unaccounted:	 0.12112294137477875 ms
sort 0.05877600610256195 ms
2842 number of unique mon objects from Dex.mod(...)
1677 number of above objects that pass gen filters
*** entire run time:			 4.6098770052194595 ms ***
*** unaccounted:			 1.4685320034623146 ms ***
```
</details>

<details>
<summary>After</summary>

```
Before
		now executing: pivot|batonpass, mod=gen8
query prep:	 52.78931099921465 ms
filtered dex size: 804
filters total:	 48.023138001561165 ms
	stat filters:	 0.11502217501401901 ms
	egg filters:	 0.10442205518484116 ms
	type filters:	 0.113189198076725 ms
	resist filters:	 0.0974508672952652 ms
	weak filters:	 0.12177606672048569 ms
	ability filters: 0.09783709049224854 ms
	forme filters:	 0.10375013947486877 ms
	prepare move filters: 1.4203269928693771 ms
		format:   	 0.2573249936103821 ms
		ruletable:	 1.0147070065140724 ms
		validator:	 0.07130900025367737 ms
		pokesource:	 0.06653899699449539 ms
		unaccounted:	 0.010446995496749878 ms
	move filters:	 44.088605120778084 ms
		get moves:	 0.02246899902820587 ms
		learn check:	 42.81362897902727 ms
		unaccounted:	 1.2525071427226067 ms
sort 0.5674260035157204 ms
1421 number of unique mon objects from Dex.mod(...)
804 number of above objects that pass gen filters
*** entire run time:			 150.75026900321245 ms ***
*** unaccounted:			 49.37039399892092 ms ***
․
		now executing: pivot, mod=gen8
query prep:	 0.7280709967017174 ms
filtered dex size: 804
filters total:	 20.988303996622562 ms
	stat filters:	 0.07573795318603516 ms
	egg filters:	 0.08135000616312027 ms
	type filters:	 0.07815118134021759 ms
	resist filters:	 0.07936599105596542 ms
	weak filters:	 0.09821894764900208 ms
	ability filters: 0.07655186951160431 ms
	forme filters:	 0.07402002066373825 ms
	prepare move filters: 0.18907099962234497 ms
		format:   	 0.1695460006594658 ms
		ruletable:	 0.0077790021896362305 ms
		validator:	 0.009063996374607086 ms
		pokesource:	 0.001828998327255249 ms
		unaccounted:	 0.0008530020713806152 ms
	move filters:	 18.556805096566677 ms
		get moves:	 0.008464999496936798 ms
		learn check:	 17.526883997023106 ms
		unaccounted:	 1.0214561000466347 ms
sort 0.38750099390745163 ms
1421 number of unique mon objects from Dex.mod(...)
804 number of above objects that pass gen filters
*** entire run time:			 23.47478500008583 ms ***
*** unaccounted:			 1.3709090128540993 ms ***
․․․․․․․․․․․
		now executing: ice, monotype
query prep:	 0.11987200379371643 ms
filtered dex size: 873
filters total:	 3.3700819984078407 ms
	stat filters:	 0.07427005469799042 ms
	egg filters:	 0.07713989913463593 ms
	type filters:	 0.3298250213265419 ms
	resist filters:	 0.07383093982934952 ms
	weak filters:	 0.07409008592367172 ms
	ability filters: 0.07394805550575256 ms
	forme filters:	 0.07313401997089386 ms
	prepare move filters: 1.386282004415989 ms
		format:   	 0.10254500061273575 ms
		ruletable:	 0.832787998020649 ms
		validator:	 0.44645099341869354 ms
		pokesource:	 0.0018340051174163818 ms
		unaccounted:	 0.002664007246494293 ms
	move filters:	 0.10970210283994675 ms
		get moves:	 0.00045800209045410156 ms
		learn check:	 0 ms
		unaccounted:	 0.10924410074949265 ms
sort 0.03980399668216705 ms
2842 number of unique mon objects from Dex.mod(...)
1677 number of above objects that pass gen filters
*** entire run time:			 5.092786997556686 ms ***
*** unaccounted:			 1.5630289986729622 ms ***
․
		now executing: ice, monotype, spe desc
query prep:	 0.09174399822950363 ms
filtered dex size: 873
filters total:	 2.9397659972310066 ms
	stat filters:	 0.07370498031377792 ms
	egg filters:	 0.07846103608608246 ms
	type filters:	 0.19768382608890533 ms
	resist filters:	 0.07355014979839325 ms
	weak filters:	 0.07839395850896835 ms
	ability filters: 0.07384713739156723 ms
	forme filters:	 0.1159859448671341 ms
	prepare move filters: 0.9995559975504875 ms
		format:   	 0.09128499776124954 ms
		ruletable:	 0.49932999908924103 ms
		validator:	 0.40650100260972977 ms
		pokesource:	 0.0012530013918876648 ms
		unaccounted:	 0.0011869966983795166 ms
	move filters:	 0.10992196947336197 ms
		get moves:	 0.0005569979548454285 ms
		learn check:	 0 ms
		unaccounted:	 0.10936497151851654 ms
sort 0.03881300240755081 ms
2842 number of unique mon objects from Dex.mod(...)
1677 number of above objects that pass gen filters
*** entire run time:			 4.067071005702019 ms ***
*** unaccounted:			 0.9967480078339577 ms ***

		now executing: ice, monotype, hp desc
query prep:	 0.06570400297641754 ms
filtered dex size: 873
filters total:	 3.096519999206066 ms
	stat filters:	 0.07612402737140656 ms
	egg filters:	 0.08095891028642654 ms
	type filters:	 0.39802011102437973 ms
	resist filters:	 0.07463895529508591 ms
	weak filters:	 0.0749899223446846 ms
	ability filters: 0.07522894442081451 ms
	forme filters:	 0.07583291828632355 ms
	prepare move filters: 0.9969040006399155 ms
		format:   	 0.0922360047698021 ms
		ruletable:	 0.5020010024309158 ms
		validator:	 0.3996400013566017 ms
		pokesource:	 0.0018050000071525574 ms
		unaccounted:	 0.0012219920754432678 ms
	move filters:	 0.11649596691131592 ms
		get moves:	 0.0005770027637481689 ms
		learn check:	 0 ms
		unaccounted:	 0.11591896414756775 ms
sort 0.040344998240470886 ms
2842 number of unique mon objects from Dex.mod(...)
1677 number of above objects that pass gen filters
*** entire run time:			 4.175630994141102 ms ***
*** unaccounted:			 0.9730619937181473 ms ***
```
</details>